+++ test/integration/test_ad_hoc.rb
@@ -216,5 +216,48 @@ class IntegrationTestAdHoc < Loofah::TestCase
         assert_nil sanitized.at_css("animate")["values"]
       end
     end
+
+    #
+    #  brought up by https://github.com/flavorjones/loofah/issues/80
+    #
+    context "comments outside html" do
+      context "bare comments" do
+        HTML = "<!-- --!><script>alert(1)</script><!-- -->"
+
+        it "Loofah.document removes the comment" do
+          sanitized = Loofah.document(HTML)
+          sanitized_html = sanitized.to_html
+          refute_match(/--/, sanitized_html)
+        end
+
+        it "Loofah.scrub_document removes the comment" do
+          sanitized = Loofah.scrub_document(HTML, :prune)
+          sanitized_html = sanitized.to_html
+          refute_match(/--/, sanitized_html)
+        end
+      end
+
+      context "doc with comments outside HTML" do
+        HTML = <<~EOF
+          <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
+              <!-- spaces ->
+          		<!-- tabs -->
+                   <!-- more spaces -->
+          <html><body><div>hello
+        EOF
+
+        it "Loofah.document removes the comment" do
+          sanitized = Loofah.document(HTML)
+          sanitized_html = sanitized.to_html
+          refute_match(/--/, sanitized_html)
+        end
+
+        it "Loofah.scrub_document removes the comment" do
+          sanitized = Loofah.scrub_document(HTML, :prune)
+          sanitized_html = sanitized.to_html
+          refute_match(/--/, sanitized_html)
+        end
+      end
+    end
   end
 end